backend/pyproject.toml
[project]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = [{ name = "Gabriel Hernandes", email = "gabriel.hernandess01@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["django (>=5.1.6,<6.0.0)", "djangorestframework (>=3.16.1,<4.0.0)", "psycopg2-binary (>=2.9.10,<3.0.0)", "celery (>=5.5.3,<6.0.0)", "redis (>=6.4.0,<7.0.0)", "djangorestframework-simplejwt (>=5.5.1,<6.0.0)", "django-cors-headers (>=4.8.0,<5.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
